Change Environment.OSVersion to use RtlGetVersion
