--- src/libraries/Common/src/Interop/Windows/Kernel32/Interop.GetVersionExW.cs
@@ -1,14 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Runtime.InteropServices;
-
-internal static partial class Interop
-{
-    internal static partial class Kernel32
-    {
-        [DllImport(Libraries.Kernel32, CharSet = CharSet.Unicode, SetLastError = true)]
-        internal static extern bool GetVersionExW(ref OSVERSIONINFOEX osvi);
-    }
-}
