+++ src/libraries/Common/src/Interop/Windows/NtDll/Interop.RtlGetVersion.cs
@@ -18,19 +18,15 @@ internal partial class Interop
             return RtlGetVersion(ref osvi);
         }
 
-        internal static unsafe string RtlGetVersion()
+        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Unicode)]
+        internal unsafe struct RTL_OSVERSIONINFOEX
         {
-            const string Version = "Microsoft Windows";
-            if (RtlGetVersionEx(out RTL_OSVERSIONINFOEX osvi) == 0)
-            {
-                return osvi.szCSDVersion[0] != '\0' ?
-                    string.Format("{0} {1}.{2}.{3} {4}", Version, osvi.dwMajorVersion, osvi.dwMinorVersion, osvi.dwBuildNumber, new string(&(osvi.szCSDVersion[0]))) :
-                    string.Format("{0} {1}.{2}.{3}", Version, osvi.dwMajorVersion, osvi.dwMinorVersion, osvi.dwBuildNumber);
-            }
-            else
-            {
-                return Version;
-            }
+            internal uint dwOSVersionInfoSize;
+            internal uint dwMajorVersion;
+            internal uint dwMinorVersion;
+            internal uint dwBuildNumber;
+            internal uint dwPlatformId;
+            internal fixed char szCSDVersion[128];
         }
     }
 }
